+++ rust-doomgeneric/src/p_plats.rs
@@ -34,9 +34,9 @@ pub const raiseToNearestAndChange: plattype_e = 3;
 pub const raiseAndChange: plattype_e = 2;
 pub const downWaitUpStay: plattype_e = 1;
 pub const perpetualRaise: plattype_e = 0;
-pub const PLATWAIT: i32 = 3 as i32;
+pub const PLATWAIT: i32 = 3;
 pub const PLATSPEED: i32 = FRACUNIT;
-pub const MAXPLATS: i32 = 30 as i32;
+pub const MAXPLATS: i32 = 30;
 pub static mut activeplats: [*mut plat_t; 30] = [::core::ptr::null::<plat_t>()
     as *mut plat_t; 30];
 pub unsafe fn T_PlatRaise(mut plat: *mut plat_t) {
